Ubuntu 25.10 Compatibility (#559)
* Update signal installation

* Update uninstall script to remove Neovim files

* Prevents keyring duplication in vscode installation

Co-Authored by @neidiom

* Move pipx installation into libraries

* Use pipx for tldr installation

* Removed Open with Alacritty option on Nautilus

* Replace Flatpak installation with direct download and installation of Obsidian

* Replace snap installation of Pinta with Flatpak installation

* Refactor Typora installation to use official repository and keyring

* Update Gum version to 0.17.0 for improved functionality

* Move gir1 installation into gnome-extensions step

* Replace Flatpak uninstallation of Obsidian with APT removal

* Replace snap removal of Pinta with Flatpak uninstallation

* Fix typo

* Simplify 1Password installation by using direct download instead of repository setup

* Update Obsidian installation script to use correct version retrieval method

Fix typo

* Use 'apt' instead of 'apt-get' for 1Password uninstallation

* Add migration script to force upgrade gum to version 0.17.0

* Refactor applications installation scripts to use a more robust method

* Remove existing Microsoft GPG key before downloading a new one

* Add fallback to reboot confirmation in desktop installer script

* Remove both keyring and sources files after installation in app-signal.sh

+++ install/desktop/app-vscode.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 
-cd /tmp
-wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg
-sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
-echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list >/dev/null
-rm -f packages.microsoft.gpg
-cd -
+if [ ! -f /etc/apt/keyrings/packages.microsoft.gpg ] || [ ! -f /usr/share/keyrings/microsoft.gpg ]; then
+  [ -f /etc/apt/keyrings/packages.microsoft.gpg ] && sudo rm /etc/apt/keyrings/packages.microsoft.gpg
+  cd /tmp
+  wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg
+  sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
+  echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list >/dev/null
+  rm -f packages.microsoft.gpg
+  cd -
+fi
 
-sudo apt update -y
+sudo apt update
 sudo apt install -y code
 
 mkdir -p ~/.config/Code/User
